% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncode.R
\name{BestCatEncode}
\alias{BestCatEncode}
\title{Detects categorical variables in a given dataset and encodes it to numeric values}
\usage{
BestCatEncode(data, dv)
}
\arguments{
\item{data}{Any dataset with atleast one categorical field that is to be encoded}

\item{dv}{Dependent variable in the given dataset}
}
\value{
Returns a list with an encoded dataset, performance metrics and a fit file to fit the test data.
}
\description{
Takes in a dataset, detects categorical variables and encodes the data using multiple encoding techniques:
}
\details{
\strong{Contrast Encoding}
\itemize{
\item Dummy Coding
\item Simple Coding
\item Deviation Coding
\item Polynomial Coding
\item Reverse Helmert Coding
\item Helmert Coding
\item Forward Difference Coding
\item Backward Difference Coding
}

\strong{Generic Encoders}
\itemize{
\item Label/Integer Encoding
\item One-Hot Encoding
\item Binary Encoding
\item Frequency Encoding
\item Hashing Encoding
\item Probability Ratio Encoding
}

\strong{Bayesian Encoders}
\itemize{
\item Mean Target Encoding
\item Leave One Out Encoding
\item Weight Of Evidence Encoding
\item James Stein Encoding
\item M-Estimator Encoding
}
Each of these 19 datasets are passed to a logistic model to choose the best possible dataset.
}
