% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodePrerequisites.R
\name{binary_encoding}
\alias{binary_encoding}
\title{Binary Encoding}
\usage{
binary_encoding(x, name)
}
\arguments{
\item{x}{A vector that needs to be encoded}

\item{name}{A string which needs to given as a field name after encoding}
}
\value{
Returns a vector with encoded data
}
\description{
Takes in a vector and encodes using Binary Encoding
}
