% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodePrerequisites.R
\name{encode_freq}
\alias{encode_freq}
\title{Frequency Encoding}
\usage{
encode_freq(x)
}
\arguments{
\item{x}{Any categorical vector which needs to be encoded}
}
\value{
Returns a list with encoded data and fit file to fit the test data
}
\description{
Takes in a vector and encodes using Frequency Encoding
}
