% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodeFit.R
\name{encode_generic_fit}
\alias{encode_generic_fit}
\title{A fit function to encode categorical data}
\usage{
encode_generic_fit(x, fit)
}
\arguments{
\item{x}{Any categorical vector which needs to be encoded}

\item{fit}{A list returned from "BestCatEncode" that is used to fit the test data.}
}
\value{
Returns the encoded data vector
}
\description{
Detects the categorical variables and treats it based on the fit file generated by train data.
}
