% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodePrerequisites.R
\name{encode_hash}
\alias{encode_hash}
\title{Hashing Encoding}
\usage{
encode_hash(df, colname)
}
\arguments{
\item{df}{A dataset with atleast one categorical field}

\item{colname}{A string with field name of categorical field}
}
\value{
Returns a list with encoded data and fit file to fit the test data
}
\description{
Takes in a vector and encodes using Hashing Encoding
}
