% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodePrerequisites.R
\name{encode_james_stein}
\alias{encode_james_stein}
\title{James Stein Encoding}
\usage{
encode_james_stein(x, y)
}
\arguments{
\item{x}{Any categorical vector which needs to be encoded}

\item{y}{Dependent variable of the dataset}
}
\value{
Returns a list with encoded data and fit file to fit the test data
}
\description{
Takes in a vector and encodes using James Stein encoding
}
