% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodePrerequisites.R
\name{encode_label}
\alias{encode_label}
\title{Label/Integer Encoding}
\usage{
encode_label(x)
}
\arguments{
\item{x}{Any categorical vector which needs to be encoded}
}
\value{
Returns a list with encoded data and fit file to fit the test data
}
\description{
Takes in a vector and encodes using Label/Integer Encoding
}
