% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodePrerequisites.R
\name{encode_m_estimator}
\alias{encode_m_estimator}
\title{M-Estimator Encoding}
\usage{
encode_m_estimator(x, y, m)
}
\arguments{
\item{x}{Any categorical vector which needs to be encoded}

\item{y}{Dependent variable of the dataset}

\item{m}{M value defaulted to 1}
}
\value{
Returns a list with encoded data and fit file to fit the test data
}
\description{
Takes in a vector and encodes using M-Estimator encoding
}
