% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodePrerequisites.R
\name{encode_onehot}
\alias{encode_onehot}
\title{One-Hot Encoding}
\usage{
encode_onehot(df, colname)
}
\arguments{
\item{df}{A dataset with atleast one categorical field}

\item{colname}{A string with field name of categorical field}
}
\value{
Returns a list with encoded data and fit file to fit the test data
}
\description{
Takes in a vector and encodes using One-Hot Encoding
}
