% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatEncodeFit.R
\name{encode_onehot_fit}
\alias{encode_onehot_fit}
\title{A fit function to encode categorical data}
\usage{
encode_onehot_fit(df, colname, fit)
}
\arguments{
\item{df}{Any dataset with atleast one categorical field}

\item{colname}{A string representing the name of the categorical field in the given dataset}

\item{fit}{A list returned from "BestCatEncode" that is used to fit the test data.}
}
\value{
Returns the encoded data vector
}
\description{
Detects the categorical variables and treats it based on the fit file generated by train data using One-Hot Encoding.
}
